# docs/api/openapi.v1.yaml
openapi: 3.0.3
info:
  title: MedTracker API
  version: v1
  description: |
    Hosted MedTracker API contract for mobile clients, API clients, and portable migration.
    API requests are limited per client IP to 300 requests in each five-minute window. GET
    household data exports are also limited to 10 per minute; POST sync batches to 30 per
    minute; GET medication lookup to 60 per minute; POST AI medication suggestions to 10 per
    minute; and GET household admin audit logs to 100 per minute. Limits use epoch-aligned
    fixed windows. The first exceeded limit rejects the request with 429 and Retry-After,
    RateLimit-Limit, RateLimit-Remaining and RateLimit-Reset response headers. Counters are
    local to each server process. Counter storage is bounded; a new client is rejected with
    429 while the store is full of live counters. Expired counters are reclaimed. Health
    checks and direct loopback clients are exempt in production. The client IP is the socket
    peer unless a deployment explicitly configures trusted proxy IPs through
    API_TRUSTED_PROXY_IPS. A configured loopback proxy is not exempt, and untrusted
    forwarding headers do not set the client IP.
    Ordinary household operations accept a valid API session, household app token,
    or mobile OAuth bearer with the medtracker scope. SMART integration grants
    authorize SMART/FHIR access, not ordinary household operations. A household
...
    last_used_at at most once every five minutes.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /households/{household_id}/medication_pause_periods:
    get:
      operationId: listMedicationPausePeriods
      tags: [Household, Schedules]
      summary: List visible medication pause history.
      description: Supply source_type and portable source_id together to select one source. Results include current and completed periods newest first by creation time, then ID, including history for retired sources. Retired sources cannot be paused or resumed.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
        - name: source_type
          in: query
          schema:
            type: string
            enum: [schedule, person_medication]
        - name: source_id
          in: query
          schema:
            $ref: '#/components/schemas/PortableId'
      responses:
        '200':
          description: Visible pause periods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationPausePeriodCollectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: createMedicationPausePeriod
      tags: [Household, Schedules]
      summary: Pause a medication source with a reason.
      description: Requires manage access. The start is server acceptance time; client timestamps are rejected with 422. An already open period is returned unchanged. No edit or backdating operation is provided.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/idempotency_key'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationPausePeriodCreateRequest'
      responses:
        '201':
          description: Pause period created or already open.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationPausePeriodResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /households/{household_id}/medication_pause_periods/{id}/resume:
    post:
      operationId: resumeMedicationPausePeriod
      tags: [Household, Schedules]
      summary: Resume the source for a recorded pause period.
      description: Requires manage access. Ends the addressed period at server acceptance time. Retrying a completed period returns it unchanged and cannot close a newer pause. The original reason, note and recording actor remain intact.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/PortableId'
        - $ref: '#/components/parameters/idempotency_key'
        - $ref: '#/components/parameters/if_match'
      responses:
        '200':
          description: Addressed completed pause period.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationPausePeriodResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Household API sessions, household app tokens, and mobile OAuth medtracker access tokens authorize ordinary household operations. SMART integration grants do not.
  parameters:
    idempotency_key:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: A key is scoped to one household for 24 hours. The same authorized account can replay a matching method, path and request through another current credential; another account or changed request receives 409. Current authentication and operation authorization are checked before replay. A replay does not repeat the domain write or its domain audit.
    household_id:
      name: household_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/NumericId'
...
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    per_page:
      name: per_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    if_match:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
...
  headers:
...
    etag:
      description: Version identifier for optimistic concurrency.
      required: true
      schema:
        type: string
    retry_after:
      description: Seconds until the request may be retried.
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
...
    BadRequest:
      description: A required request parameter is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: The bearer credential is absent, invalid, expired, revoked, or no longer active.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The credential cannot perform the operation in the routed household or person scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: The resource does not exist inside the authorized scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationFailed:
      description: The submitted attributes or filters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: The resource or idempotency state conflicts with the requested mutation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: The request exceeded an API rate limit.
      headers:
        Retry-After:
          $ref: '#/components/headers/retry_after'
        RateLimit-Limit:
          description: Maximum requests allowed by the matched limit.
          schema:
            type: integer
            minimum: 1
        RateLimit-Remaining:
          description: Requests remaining in the matched window.
          schema:
            type: integer
            enum: [0]
        RateLimit-Reset:
          description: Unix timestamp in seconds when the matched window resets.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitErrorEnvelope'
  schemas:
    NumericId:
      type: integer
      minimum: 1
    PortableId:
      type: string
      format: uuid
...
    PaginationMeta:
      type: object
      additionalProperties: false
      required:
        - page
        - per_page
        - total_count
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
        total_count:
          type: integer
          minimum: 0
...
    MedicationPausePeriod:
      type: object
      additionalProperties: false
      required: [id, portable_id, source_type, source_id, reason, note, legacy_context, started_at, ended_at, recorded_by_membership_id, resumed_by_membership_id, recorded_by_name, resumed_by_name, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/PortableId'
        portable_id:
          $ref: '#/components/schemas/PortableId'
        source_type:
          type: string
          enum: [schedule, person_medication]
        source_id:
          $ref: '#/components/schemas/PortableId'
        reason:
          type: string
          enum: [out_of_supply, temporarily_not_needed, clinician_advice, side_effects, other, reason_not_recorded]
        note:
          type: string
          nullable: true
        legacy_context:
          type: boolean
        started_at:
          type: string
          format: date-time
          nullable: true
          description: Unknown for migrated pauses; new requests use server acceptance time.
        ended_at:
          type: string
          format: date-time
          nullable: true
        recorded_by_membership_id:
          type: string
          nullable: true
        resumed_by_membership_id:
          type: string
          nullable: true
        recorded_by_name:
          type: string
          nullable: true
        resumed_by_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MedicationPausePeriodCreateRequest:
      type: object
      additionalProperties: false
      required: [medication_pause_period]
      properties:
        medication_pause_period:
          type: object
          additionalProperties: false
          required: [source_type, source_id, reason]
          properties:
            source_type:
              type: string
              enum: [schedule, person_medication]
            source_id:
              $ref: '#/components/schemas/PortableId'
            reason:
              type: string
              enum: [out_of_supply, temporarily_not_needed, clinician_advice, side_effects, other]
            note:
              type: string
              nullable: true
    MedicationPausePeriodResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/MedicationPausePeriod'
    MedicationPausePeriodCollectionResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MedicationPausePeriod'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    ValidationErrors:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1
        errors:
          $ref: '#/components/schemas/ValidationErrors'
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
...
    RateLimitError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - rate_limited
        message:
          type: string
          minLength: 1
    RateLimitErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/RateLimitError'
...